[ 2609057 ] Co Product functionality http://sourceforge.net/tracker/index.php?func=detail&aid=2609057&group_id=176962&atid=934929
Implement CostAllocationPerc functionality: in a lot of cases cost absorption of the co-product is not the same as bom qty. In this case we need in bom line another field which is the cost absorption percent (from 0 to 100) which specify the absorption percent.

--HG--
branch : development
extra : convert_revision : svn%3Aca97e3cf-264c-4fc5-ac59-d2cac9e95311/trunk%4010983